# docs/swagger.yaml
basePath: /
definitions:
  handler.PagedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/store.UserRank'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 150
        type: integer
    type: object
  store.ScoreEvent:
    properties:
      rank:
        example: 1
        type: integer
      score:
        example: 981
        type: integer
      username:
        example: alice
        type: string
    type: object
  store.UserRank:
    properties:
      rank:
        example: 1
        type: integer
      score:
        example: 980
        type: integer
      username:
        example: alice
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: moustafa_algammal@yahoo.com
    name: Moustafa Elgammal
  description: Internal leaderboard service. Designed to sit behind a game service
    that handles player authentication, session management, anti-cheat, and per-player
    rate limiting. All requests must include the X-Internal-Token header when INTERNAL_API_KEY
    is configured.
  title: Real-Time Gaming Leaderboard API
  version: "1.0"
paths:
  /v1/scores:
    get:
      description: Returns a page of players for the current calendar month, ordered
        by score descending. Defaults to page 1 with page_size equal to the TOP_N
        configuration value.
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of entries per page (1-100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PagedResponse'
        "400":
          description: Invalid page or page_size
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - InternalToken: []
      summary: Get top scores (paginated)
      tags:
      - leaderboard
  /v1/scores/{username}:
    get:
      description: Returns the player along with up to USER_NEIGHBORHOOD players ranked
        immediately above and below them in the current month's leaderboard.
      parameters:
      - description: Player username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.UserRank'
            type: array
        "404":
          description: Player not on leaderboard
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - InternalToken: []
      summary: Get player rank and neighbors
      tags:
      - leaderboard
    post:
      description: Adds 1 point to the player's score for the current month. Called
        by the game service after it has authenticated the player and validated the
        score event. Creates the player entry if it does not exist.
      parameters:
      - description: Player username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing username
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: MySQL or Redis write failure
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - InternalToken: []
      summary: Increment player score
      tags:
      - leaderboard
  /v1/scores/stream:
    get:
      description: Opens a Server-Sent Events stream. Each `score-update` event carries
        the player's username, new score, and updated rank. The stream stays open
        until the client disconnects. Connect via `EventSource` in a browser or any
        HTTP client that supports `text/event-stream`.
      produces:
      - text/event-stream
      responses:
        "200":
          description: Continuous stream of score-update SSE events
          schema:
            $ref: '#/definitions/store.ScoreEvent'
      security:
      - InternalToken: []
      summary: Stream live score updates (SSE)
      tags:
      - leaderboard
securityDefinitions:
  InternalToken:
    description: Shared secret set via INTERNAL_API_KEY env var. Leave empty to disable
      (local dev).
    in: header
    name: X-Internal-Token
    type: apiKey
swagger: "2.0"